+++ test/Fixture/Methods/InvokeParentHookMethodRule/Codeception/Cest/ConcreteCestCaseNotInvokingEmptyParentHookMethodsCestHook.php
@@ -1,0 +1,26 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Ergebnis\PHPStan\Rules\Test\Fixture\Methods\InvokeParentHookMethodRule\Codeception\Cest;
+
+class ConcreteCestCaseNotInvokingEmptyParentHookMethodsCestHook extends AbstractCestCaseWithEmptyHookMethods
+{
+    public function _before(): void
+    {
+        self::setUpSomethingElse();
+    }
+
+    public function _after(): void
+    {
+        self::tearDownSomethingElse();
+    }
+
+    private static function setUpSomethingElse(): void
+    {
+    }
+
+    private static function tearDownSomethingElse(): void
+    {
+    }
+}