+++ test/Util/Helper.php
@@ -13,10 +13,30 @@ declare(strict_types=1);
 
 namespace Ergebnis\PHPStan\Rules\Test\Util;
 
+use Faker\Factory;
+use Faker\Generator;
 use Symfony\Component\Finder;
 
 trait Helper
 {
+    final protected static function faker(string $locale = 'en_US'): Generator
+    {
+        /**
+         * @var array<string, Generator> $fakers
+         */
+        static $fakers = [];
+
+        if (!\array_key_exists($locale, $fakers)) {
+            $faker = Factory::create($locale);
+
+            $faker->seed(9001);
+
+            $fakers[$locale] = $faker;
+        }
+
+        return $fakers[$locale];
+    }
+
     /**
      * @return list<string>
      */